refactor(db): restructure database service files and models for improved organization and maintainability
